Move the helper function used in fieldmanager test files to another go file for easy external reuse
