+++ staging/src/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/skipnonapplied_test.go
@@ -23,23 +23,9 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"sigs.k8s.io/yaml"
 )
-
-type fakeObjectCreater struct {
-	gvk schema.GroupVersionKind
-}
-
-var _ runtime.ObjectCreater = &fakeObjectCreater{}
-
-func (f *fakeObjectCreater) New(_ schema.GroupVersionKind) (runtime.Object, error) {
-	u := unstructured.Unstructured{Object: map[string]interface{}{}}
-	u.SetAPIVersion(f.gvk.GroupVersion().String())
-	u.SetKind(f.gvk.Kind)
-	return &u, nil
-}
 
 func TestNoUpdateBeforeFirstApply(t *testing.T) {
 	f := NewTestFieldManager(schema.FromAPIVersionAndKind("v1", "Pod"), "", func(m Manager) Manager {
